requirements.txt
streamlit
pandas
pymysql
python-dotenv
plotly